backend/requirements.txt
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django==5.0.4
django-appconf==1.0.6
django-cors-headers==4.4.0
django-cryptography==1.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
mysqlclient==2.2.4
python-dotenv==1.0.1
redis==5.0.8
requests==2.31.0
uuid==1.30
